requirements.txt
discord.py[voice]>=1.5.1
googletrans==4.0.0-rc1
beautifulsoup4==4.10.0
requests==2.27.1